resource: use 32-bit inidices internally

+++ src/resource/mesh.h
@@ -37,11 +37,11 @@ struct Geometry
 	Array<f32> _tangents;
 	Array<f32> _binormals;
 
-	Array<u16> _position_indices;
-	Array<u16> _normal_indices;
-	Array<u16> _uv_indices;
-	Array<u16> _tangent_indices;
-	Array<u16> _binormal_indices;
+	Array<u32> _position_indices;
+	Array<u32> _normal_indices;
+	Array<u32> _uv_indices;
+	Array<u32> _tangent_indices;
+	Array<u32> _binormal_indices;
 
 	Array<char> _vertex_buffer;
 	Array<u16> _index_buffer;
